GH-38090: [C++][Emscripten] ipc: Suppress shorten-64-to-32 warnings
We need explicit cast to use `int64_t` for `size_t` on Emscripten.

Explicit casts.

+++ cpp/src/arrow/ipc/metadata_internal.cc
@@ -464,7 +464,7 @@ static KeyValueOffset AppendKeyValue(FBB& fbb, const std::string& key,
 
 static void AppendKeyValueMetadata(FBB& fbb, const KeyValueMetadata& metadata,
                                    std::vector<KeyValueOffset>* key_values) {
-  key_values->reserve(metadata.size());
+  key_values->reserve(static_cast<size_t>(metadata.size()));
   for (int i = 0; i < metadata.size(); ++i) {
     key_values->push_back(AppendKeyValue(fbb, metadata.key(i), metadata.value(i)));
   }
